fix: move forum post answer component
