+++ client/app/bundles/course/assessment/submission/components/answers/ForumPostResponse/ForumPostOption.jsx
@@ -1,12 +1,9 @@
 import { Component } from 'react';
 import { blueGrey, green } from '@mui/material/colors';
-
 import PropTypes from 'prop-types';
 import { injectIntl } from 'react-intl';
-
 import { postPackShape } from 'course/assessment/submission/propTypes';
-import ForumPost, { translations } from 'course/forum/components/ForumPost';
-
+import ForumPost, { translations } from './ForumPost';
 import ParentPost from './ParentPost';
 
 const styles = {
